Revert "Add --no-infer option to evaluate.py"
This reverts commit c43cf52c66767677ad47c46571774484bdf07f84.

Commit c43cf52c66767677ad47c46571774484bdf07f84 did not do what it claimed to:
it did not in fact enable calculation of the run time of inference.  That is
because post-inference processing time was included in the time difference with
and without the --no-infer flag.

Even if it had worked as intended, it was a clumsy and inefficient approach to
measurement as it required running the program multiple times and comparing the
results.

An effective and more convenient way to measure inference time is simply to do
it in code unconditionally and print the result.  😂